test(ssm): split gated delta net unit tests
